+++ packages/app-headless-cms-common/src/Fields/useBind.tsx
@@ -1,11 +1,12 @@
-import React, { useRef, useCallback, cloneElement } from "react";
+import React, { useRef, cloneElement } from "react";
 import { Validator } from "@webiny/validation/types";
 import { useForm } from "@webiny/form";
 import { createValidators } from "~/createValidators";
 import { BindComponent, CmsModelField } from "~/types";
+import { useModelField } from "~/ModelFieldProvider";
+import { createValidationScope } from "~/createValidationScope";
 
 interface UseBindProps {
-    field: CmsModelField;
     Bind: BindComponent;
 }
 
@@ -31,130 +32,127 @@ export interface GetBindCallable {
 
 const emptyValidators: Validator[] = [];
 
-export function useBind({ Bind, field }: UseBindProps) {
+export function useBind({ Bind }: UseBindProps) {
+    const { field } = useModelField();
     const memoizedBindComponents = useRef<Record<string, BindComponent>>({});
     const cacheKey = createFieldCacheKey(field);
     const form = useForm();
 
-    return useCallback(
-        (index = -1) => {
-            const { parentName } = Bind;
+    return (index = -1) => {
+        const { parentName } = Bind;
 
-            // If there's a parent name assigned to the given Bind component, we need to include it in the new field "name".
-            // This allows us to have nested fields (like "object" field with nested properties)
-            const name = [parentName, field.fieldId, index >= 0 ? index : undefined]
-                .filter(v => v !== undefined)
-                .join(".");
+        // If there's a parent name assigned to the given Bind component, we need to include it in the new field "name".
+        // This allows us to have nested fields (like "object" field with nested properties)
+        const name = [parentName, field.fieldId, index >= 0 ? index : undefined]
+            .filter(v => v !== undefined)
+            .join(".");
 
-            const componentId = `${name};${cacheKey}`;
+        const componentId = `${name};${cacheKey}`;
 
-            if (memoizedBindComponents.current[componentId]) {
-                return memoizedBindComponents.current[componentId];
-            }
-
-            const validators = createValidators(field, field.validation || emptyValidators);
-            const listValidators = createValidators(field, field.listValidation || emptyValidators);
-            const isMultipleValues = index === -1 && field.multipleValues;
-            const inputValidators = isMultipleValues ? listValidators : validators;
-
-            // We only use default values for single-value fields.
-            const defaultValueFromSettings = !isMultipleValues
-                ? field.settings?.defaultValue
-                : null;
-
-            memoizedBindComponents.current[componentId] = function UseBind(params: UseBindParams) {
-                const {
-                    name: childName,
-                    validators: childValidators,
-                    children,
-                    defaultValue = defaultValueFromSettings
-                } = params;
-
-                return (
-                    <Bind
-                        name={childName || name}
-                        validators={childValidators || inputValidators}
-                        defaultValue={defaultValue ?? null}
-                    >
-                        {bind => {
-                            // Multiple-values functions below.
-                            const props = { ...bind };
-                            if (field.multipleValues && index === -1) {
-                                props.appendValue = (newValue: any, index?: number) => {
-                                    const currentValue = bind.value || [];
-                                    const newIndex = index ?? currentValue.length;
-
-                                    bind.onChange([
-                                        ...currentValue.slice(0, newIndex),
-                                        newValue,
-                                        ...currentValue.slice(newIndex)
-                                    ]);
-                                };
-                                props.prependValue = (newValue: any) => {
-                                    bind.onChange([newValue, ...(bind.value || [])]);
-                                };
-                                props.appendValues = (newValues: any[]) => {
-                                    bind.onChange([...(bind.value || []), ...newValues]);
-                                };
-
-                                props.removeValue = (index: number) => {
-                                    if (index < 0) {
-                                        return;
-                                    }
-
-                                    const value = [
-                                        ...bind.value.slice(0, index),
-                                        ...bind.value.slice(index + 1)
-                                    ];
-
-                                    bind.onChange(value.length === 0 ? null : value);
-
-                                    // To make sure the field is still valid, we must trigger validation.
-                                    form.validateInput(field.fieldId);
-                                };
-
-                                props.moveValueUp = (index: number) => {
-                                    if (index <= 0) {
-                                        return;
-                                    }
-
-                                    const value = [...bind.value];
-                                    value.splice(index, 1);
-                                    value.splice(index - 1, 0, bind.value[index]);
-
-                                    bind.onChange(value);
-                                };
-
-                                props.moveValueDown = (index: number) => {
-                                    if (index >= bind.value.length) {
-                                        return;
-                                    }
-
-                                    const value = [...bind.value];
-                                    value.splice(index, 1);
-                                    value.splice(index + 1, 0, bind.value[index]);
-
-                                    bind.onChange(value);
-                                };
-                            }
-
-                            const element =
-                                typeof children === "function"
-                                    ? children(props)
-                                    : cloneElement(children, props);
-
-                            return element;
-                        }}
-                    </Bind>
-                );
-            } as BindComponent;
-
-            // We need to keep track of current field name, to support nested fields.
-            memoizedBindComponents.current[componentId].parentName = name;
-            memoizedBindComponents.current[componentId].displayName = `Bind<${name}>`;
-
+        if (memoizedBindComponents.current[componentId]) {
             return memoizedBindComponents.current[componentId];
-        },
-        [field.fieldId, cacheKey]
-    );
+        }
+
+        const validators = createValidators(field, field.validation || emptyValidators);
+        const listValidators = createValidators(field, field.listValidation || emptyValidators);
+        const isMultipleValues = index === -1 && field.multipleValues;
+        const inputValidators = isMultipleValues ? listValidators : validators;
+
+        // We only use default values for single-value fields.
+        const defaultValueFromSettings = !isMultipleValues ? field.settings?.defaultValue : null;
+
+        memoizedBindComponents.current[componentId] = function UseBind(params: UseBindParams) {
+            const {
+                name: childName,
+                validators: childValidators,
+                children,
+                defaultValue = defaultValueFromSettings
+            } = params;
+
+            const { field } = useModelField();
+
+            return (
+                <Bind
+                    name={childName || name}
+                    validators={childValidators || inputValidators}
+                    defaultValue={defaultValue ?? null}
+                    context={{ field }}
+                >
+                    {bind => {
+                        // Multiple-values functions below.
+                        const props = { ...bind };
+                        if (field.multipleValues && index === -1) {
+                            props.appendValue = (newValue: any, index?: number) => {
+                                const currentValue = bind.value || [];
+                                const newIndex = index ?? currentValue.length;
+
+                                bind.onChange([
+                                    ...currentValue.slice(0, newIndex),
+                                    newValue,
+                                    ...currentValue.slice(newIndex)
+                                ]);
+                            };
+                            props.prependValue = (newValue: any) => {
+                                bind.onChange([newValue, ...(bind.value || [])]);
+                            };
+                            props.appendValues = (newValues: any[]) => {
+                                bind.onChange([...(bind.value || []), ...newValues]);
+                            };
+
+                            props.removeValue = (index: number) => {
+                                if (index < 0) {
+                                    return;
+                                }
+
+                                const value = [
+                                    ...bind.value.slice(0, index),
+                                    ...bind.value.slice(index + 1)
+                                ];
+
+                                bind.onChange(value.length === 0 ? null : value);
+
+                                // To make sure the field is still valid, we must trigger validation.
+                                form.validateInput(field.fieldId);
+                            };
+
+                            props.moveValueUp = (index: number) => {
+                                if (index <= 0) {
+                                    return;
+                                }
+
+                                const value = [...bind.value];
+                                value.splice(index, 1);
+                                value.splice(index - 1, 0, bind.value[index]);
+
+                                bind.onChange(value);
+                            };
+
+                            props.moveValueDown = (index: number) => {
+                                if (index >= bind.value.length) {
+                                    return;
+                                }
+
+                                const value = [...bind.value];
+                                value.splice(index, 1);
+                                value.splice(index + 1, 0, bind.value[index]);
+
+                                bind.onChange(value);
+                            };
+                        }
+
+                        return typeof children === "function"
+                            ? children(props)
+                            : cloneElement(children, props);
+                    }}
+                </Bind>
+            );
+        } as BindComponent;
+
+        // We need to keep track of current field name, to support nested fields.
+        memoizedBindComponents.current[componentId].parentName = name;
+        memoizedBindComponents.current[componentId].displayName = `Bind<${name}>`;
+        memoizedBindComponents.current[componentId].ValidationScope = createValidationScope(name);
+
+        return memoizedBindComponents.current[componentId];
+    };
 }
